[feat] usecase , face 기능 분리
