+++ src/main/java/kr/hhplus/be/server/concert/controller/ConcertController.java
@@ -1,8 +1,8 @@
 package kr.hhplus.be.server.concert.controller;
 
 import kr.hhplus.be.server.concert.application.dtos.*;
+import kr.hhplus.be.server.reservation.application.facade.ReservationFacade;
 import kr.hhplus.be.server.concert.application.service.ConcertService;
-import kr.hhplus.be.server.concert.application.userCase.ReserveUseCase;
 import kr.hhplus.be.server.concert.domain.Concert;
 import kr.hhplus.be.server.config.ui.Response;
 import kr.hhplus.be.server.seat.application.service.SeatService;
@@ -16,7 +16,7 @@ import org.springframework.web.bind.annotation.*;
 public class ConcertController {
     private final ConcertService concertService;
     private final SeatService seatService;
-    private final ReserveUseCase reserveUseCase;
+    private final ReservationFacade reservationFacade;
 
     //콘서트 조회
     @GetMapping("/find")
@@ -37,7 +37,8 @@ public class ConcertController {
     //콘서트 자리 선택
     @PostMapping("/{concertId}/seats/choice")
     public Response<ConcertSeatInfoResponse>choiceSeat(@PathVariable("concertId")Long concertId,@RequestParam("seatId") Long seatId ,@RequestParam("userId") Long userId){
-        Seat seat = reserveUseCase.choiceSeatAndReserve(concertId, seatId,userId);
+//        Seat seat = reserveUseCase.choiceSeatAndReserve(concertId, seatId,userId);
+        Seat seat = reservationFacade.choiceSeatAndReserveWithLock(concertId, seatId, userId);
         ConcertSeatInfoResponse response = ConcertSeatInfoResponse.from(seat);
         return Response.ok(response);
     }
