+++ src/main/java/kr/hhplus/be/server/user/application/service/UserService.java
@@ -13,7 +13,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-@Transactional
 public class UserService {
     private final UserRepository userRepository;
     private final UserModelMapper userModelMapper;
